Remove generic from Base.ViewController
